+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## devel
 
+### Improvements
+#### Modules:
+  - zabbix_action - fixed error on changed API fields for Zabbix 5.0
+
 ## 0.2.0
 
 ### New roles: